Make datatable aware of the data column indexes
This commit will add a field named ColumnsMap to the DataTable struct.
This field is a map of type map[int]int.  It maps the ColumnIndex for a
Table definition into the actual index for each data row array.

The purpose of this map is to properly keep track of optional columns
when they are not present for a peer.  For instance, for a peer of type
Shinken, optional columns of type Naemon won't be present.  For a given
Livestatus table, there will be less columns in the datatable data
representation than the combined sum of dynamic and static columns
declared in the table definition.

If the value for a datatable column is retrieved using the table index,
and the column appears after at least an optional column that has been
stripped, there is a chance that an offset error will happen.