Fix offset errors when accessing datatable data
This commit will update the (*Peer).gatherResultsRow function so that
when accessing a particular column of a row in the datatable data, it
first uses the datatable columns map to make sure that it's accessing
the correct column.

The rationale for this commit is similar to the one used in the previous
commit.  Given a Livestatus table, if an optional column is not present
in a peer because the flags don't match, those columns won't be present
in the datatable data.  However, because currently gatherResultsRow uses
the table (not the datatable) indexes for accessing a particular cell
when computing the result, there is a risk of causing offset errors if
the column to retrieve appears after a column that has been skipped.