Make peers aware of the columns for the DataTables (#50)
* Make datatable aware of the data column indexes

This commit will add a field named ColumnsMap to the DataTable struct.
This field is a map of type map[int]int.  It maps the ColumnIndex for a
Table definition into the actual index for each data row array.

The purpose of this map is to properly keep track of optional columns
when they are not present for a peer.  For instance, for a peer of type
Shinken, optional columns of type Naemon won't be present.  For a given
Livestatus table, there will be less columns in the datatable data
representation than the combined sum of dynamic and static columns
declared in the table definition.

If the value for a datatable column is retrieved using the table index,
and the column appears after at least an optional column that has been
stripped, there is a chance that an offset error will happen.

* Fix offset errors when accessing datatable data

This commit will update the (*Peer).gatherResultsRow function so that
when accessing a particular column of a row in the datatable data, it
first uses the datatable columns map to make sure that it's accessing
the correct column.

The rationale for this commit is similar to the one used in the previous
commit.  Given a Livestatus table, if an optional column is not present
in a peer because the flags don't match, those columns won't be present
in the datatable data.  However, because currently gatherResultsRow uses
the table (not the datatable) indexes for accessing a particular cell
when computing the result, there is a risk of causing offset errors if
the column to retrieve appears after a column that has been skipped.

* Initialises columnsMap on virtual tables

It seems I misunderstood this branch. Of course, if virtual tables do
not initialise the columns map, then requests made to that table cannot
succeed.

This commit moves the initialization for the columns map before these
branchs, and then sets the map in all circumstances.