Use more of the test dataset
Previously, when generating the host and service dataset, only the last
host/service defined in the test data would be used.

This commit ensures that we use the actual dataset, and if more hosts or
services than defined are needed, then we duplicate the last defined
test object (as before).

Given that the test dataset used have changed, a bunch of the expected
values in the tests needed to be adjusted.

Signed-off-by: Jacob Hansen <jhansen@op5.com>