Change servicegroups members to InterfaceList
The members column of servicegroups where previously defined as a
stringListCol. This resulted in an incorrect format, where it would be
impossible to distinguish the hostname from the service description if
either had a space in it.

Livestatus would output the member as: ["monitor","System load average"]
whereas it would LMD would output it as: "[monitor System load average]"

By setting the servicegroup members column to InterfaceListCol we
correct the output to be in line with Livestatus.

Signed-off-by: Jacob Hansen <jhansen@op5.com>