Add support for the AuthUser header
This commit adds support for setting the AuthUser header option. The
option is used for limiting the rows returned, based on whether the
username set in the AuthUser header is a contact on the objects
requested.

The two settings ServieAuthorization and GroupAuthorization allows some
customization of the behaviour, and reflects the options present in
livestatus.

Signed-off-by: Jacob Hansen <jhansen@op5.com>