Add support for the AuthUser header (#71)
* Also use first host of test data

When more than one host was present in the testdata, the first host
would be skipped, and only the second would be used. This commit ensures
that the first host in the testdata is used as well.

It worked before before, as last variable would be set to 0.

Signed-off-by: Jacob Hansen <jhansen@op5.com>

* Add support for the AuthUser header

This commit adds support for setting the AuthUser header option. The
option is used for limiting the rows returned, based on whether the
username set in the AuthUser header is a contact on the objects
requested.

The two settings ServieAuthorization and GroupAuthorization allows some
customization of the behaviour, and reflects the options present in
livestatus.

Signed-off-by: Jacob Hansen <jhansen@op5.com>