Correctly handle negate header in filter groups
Previously if the negate header was included, the whole filter result
would be negated. However in some queries only part of the filter should
be negated, for example:

GET hosts
Filter: alias ~ test
Filter: state = 0
Negate:
And: 2

In this case, we would negate the full query, whereas in livewstatus
only the state filter would be negated.

This commit brings the behaviour in line with Livestatus.

Signed-off-by: Jacob Hansen <jhansen@itrsgroup.com>