Fix missing comments detection in delta update
When fetching a list of IDs from the source, the IDs were incorrectly
transformed (with sprintf), marking all comments as missing.

Further, this also caused all comments to be deleted, so we would
essential do a full update every time instead of a delta update.

With this commit we convert to int64 and then use `%d` instead of `%v`
when getting the ID string.

Signed-off-by: Jacob Hansen <jhansen@op5.com>