Fix travis CI script
Don't fail the CI job for pull requests.
When the job is a "pull_request", the statement fails as the
`[ $TRAVIS_EVENT_TYPE != "pull_request" ]` part exits with 1.
This doesn't look as nice, but an if-statement doesn't produce non-zero
exit even if the test fails, so it should work as intended here.